Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU) conversion for LDH

This function fits a 5-parameter logistic standard curve to the dilutions of the positive controls for each protein and converts the MFI values into relative antibody units (RAU).

Usage

MFItoRAU_LDH(sero_data, plate_list, std_point = "PvLDH", file_path = NULL)

Arguments

sero_data

Output from `readSeroData()` or `readSeroData()`.

plate_list

Output from `readPlateLayout()`.

std_point

Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve, "PvLDH" for LDH specific curve. Default = "PvLDH".

file_path

A file path to write the .csv final file. Default: Current working directory.

Value

A data frame containing the MFI and RAU Dilution values for each sample

Author

Connie Li Wai Suen, Caitlin Bourke, Dionne Argyropoulos

Examples

# \donttest{
# Example demonstrating multi-plate processing workflow.
# These files are included in the SeroTrackR package under inst/extdata.

your_raw_data <- system.file("extdata", "example_BioPlex_PvLDH_plate1.xlsx", package = "SeroTrackR")

your_plate_layout <- system.file(
  "extdata", "example_platelayout_1.xlsx",
  package = "SeroTrackR"
)

# Read in raw BioPlex data
sero_data <- readSeroData(
  raw_data = your_raw_data,
  platform = "bioplex"
)
#> PASS: File example_bioplex_pvldh_plate1.xlsx successfully validated.

# Read matching plate layout
plate_list <- readPlateLayout(
  plate_layout = your_plate_layout,
  sero_data = sero_data
)
#> Plate layouts correctly identified!

# Run MFI to RAU conversion
mfi_outputs <- MFItoRAU_LDH(
  sero_data = sero_data,
  plate_list = plate_list
)

# View All Outputs
mfi_outputs
#>    SampleID Location.2  Location Sample  Plate PvLDH-ST_MFI PvLDH-ST_Dilution
#> 1    ABC013         B1  13(1,B1)    X13 plate1         54.5         331.46776
#> 2    ABC014         B2  14(1,B2)    X14 plate1         28.5          57.56885
#> 3    ABC015         B3  15(1,B3)    X15 plate1         33.5          99.51020
#> 4    ABC016         B4  16(1,B4)    X16 plate1         32.0          86.11171
#> 5    ABC017         B5  17(1,B5)    X17 plate1         33.0          94.97366
#> 6    ABC018         B6  18(1,B6)    X18 plate1         30.0          69.29915
#> 7    ABC019         B7  19(1,B7)    X19 plate1      15646.5      168506.01492
#> 8    ABC020         B8  20(1,B8)    X20 plate1         40.0         163.90899
#> 9    ABC021         B9  21(1,B9)    X21 plate1         46.5         235.95779
#> 10   ABC022        B10 22(1,B10)    X22 plate1         32.0          86.11171
#> 11   ABC023        B11 23(1,B11)    X23 plate1         33.5          99.51020
#> 12   ABC024        B12 24(1,B12)    X24 plate1        114.0        1110.14041
#> 13   ABC025         C1  25(1,C1)    X25 plate1         29.0          61.38969
#> 14   ABC026         C2  26(1,C2)    X26 plate1         48.5         259.24703
#> 15   ABC027         C3  27(1,C3)    X27 plate1         41.0         174.57172
#> 16   ABC028         C4  28(1,C4)    X28 plate1         52.0         300.99301
#> 17   ABC029         C5  29(1,C5)    X29 plate1         29.0          61.38969
#> 18   ABC030         C6  30(1,C6)    X30 plate1         28.5          57.56885
#> 19   ABC031         C7  31(1,C7)    X31 plate1         38.0         143.13049
#> 20   ABC032         C8  32(1,C8)    X32 plate1         27.5          50.20976
#> 21   ABC033         C9  33(1,C9)    X33 plate1         40.0         163.90899
#> 22   ABC034        C10 34(1,C10)    X34 plate1         23.0          22.40309
#> 23   ABC035        C11 35(1,C11)    X35 plate1         32.0          86.11171
#> 24   ABC036        C12 36(1,C12)    X36 plate1         26.0          39.92476
#> 25   ABC037         D1  37(1,D1)    X37 plate1         32.5          90.50674
#> 26   ABC038         D2  38(1,D2)    X38 plate1         34.0         104.11415
#> 27   ABC039         D3  39(1,D3)    X39 plate1         25.5          36.71057
#> 28   ABC040         D4  40(1,D4)    X40 plate1         35.0         113.51578
#> 29   ABC041         D5  41(1,D5)    X41 plate1         73.5         574.17503
#> 30   ABC042         D6  42(1,D6)    X42 plate1         22.0          17.57084
#> 31   ABC043         D7  43(1,D7)    X43 plate1         28.5          57.56885
#> 32   ABC044         D8  44(1,D8)    X44 plate1         35.5         118.30941
#> 33   ABC045         D9  45(1,D9)    X45 plate1         24.0          27.75898
#> 34   ABC046        D10 46(1,D10)    X46 plate1         98.0         898.35665
#> 35   ABC047        D11 47(1,D11)    X47 plate1         26.0          39.92476
#> 36   ABC048        D12 48(1,D12)    X48 plate1         33.0          94.97366
#> 37   ABC049         E1  49(1,E1)    X49 plate1         25.5          36.71057
#> 38   ABC050         E2  50(1,E2)    X50 plate1         32.0          86.11171
#> 39   ABC051         E3  51(1,E3)    X51 plate1         81.5         679.50647
#> 40   ABC052         E4  52(1,E4)    X52 plate1         29.5          65.30075
#> 41   ABC053         E5  53(1,E5)    X53 plate1         25.0          33.60918
#> 42   ABC054         E6  54(1,E6)    X54 plate1         42.0         185.40211
#> 43   ABC055         E7  55(1,E7)    X55 plate1         46.5         235.95779
#> 44   ABC056         E8  56(1,E8)    X56 plate1         29.0          61.38969
#> 45   ABC057         E9  57(1,E9)    X57 plate1         24.0          27.75898
#> 46   ABC058        E10 58(1,E10)    X58 plate1         37.0         133.03862
#> 47   ABC059        E11 59(1,E11)    X59 plate1         37.5         138.05846
#> 48   ABC060        E12 60(1,E12)    X60 plate1         32.5          90.50674
#> 49   ABC061         F1  61(1,F1)    X61 plate1         25.5          36.71057
#> 50   ABC062         F2  62(1,F2)    X62 plate1         33.5          99.51020
#> 51   ABC063         F3  63(1,F3)    X63 plate1         31.0          77.54691
#> 52   ABC064         F4  64(1,F4)    X64 plate1         36.5         128.07266
#> 53   ABC065         F5  65(1,F5)    X65 plate1         22.0          17.57084
#> 54   ABC066         F6  66(1,F6)    X66 plate1         39.0         153.42474
#> 55   ABC067         F7  67(1,F7)    X67 plate1         37.5         138.05846
#> 56   ABC068         F8  68(1,F8)    X68 plate1         36.0         123.16233
#> 57   ABC069         F9  69(1,F9)    X69 plate1         22.0          17.57084
#> 58   ABC070        F10 70(1,F10)    X70 plate1         30.0          69.29915
#> 59   ABC071        F11 71(1,F11)    X71 plate1         32.0          86.11171
#> 60   ABC072        F12 72(1,F12)    X72 plate1         26.0          39.92476
#> 61   ABC073         G1  73(1,G1)    X73 plate1         25.5          36.71057
#> 62   ABC074         G2  74(1,G2)    X74 plate1         28.5          57.56885
#> 63   ABC075         G3  75(1,G3)    X75 plate1         31.0          77.54691
#> 64   ABC076         G4  76(1,G4)    X76 plate1         30.0          69.29915
#> 65   ABC077         G5  77(1,G5)    X77 plate1         27.0          46.67771
#> 66   ABC078         G6  78(1,G6)    X78 plate1         33.0          94.97366
#> 67   ABC079         G7  79(1,G7)    X79 plate1         27.5          50.20976
#> 68   ABC080         G8  80(1,G8)    X80 plate1         36.0         123.16233
#> 69   ABC081         G9  81(1,G9)    X81 plate1         22.0          17.57084
#> 70   ABC082        G10 82(1,G10)    X82 plate1         30.0          69.29915
#> 71   ABC083        G11 83(1,G11)    X83 plate1         24.5          30.62411
#> 72   ABC084        G12 84(1,G12)    X84 plate1         31.0          77.54691
#> 73   ABC085         H1  85(1,H1)    X85 plate1         26.5          43.24827
#> 74   ABC086         H2  86(1,H2)    X86 plate1         32.0          86.11171
#> 75   ABC087         H3  87(1,H3)    X87 plate1         26.0          39.92476
#> 76   ABC088         H4  88(1,H4)    X88 plate1         35.0         113.51578
#> 77   ABC089         H5  89(1,H5)    X89 plate1         24.0          27.75898
#> 78   ABC090         H6  90(1,H6)    X90 plate1         64.0         450.94228
#> 79   ABC091         H7  91(1,H7)    X91 plate1         33.5          99.51020
#> 80   ABC092         H8  92(1,H8)    X92 plate1         31.5          81.79096
#> 81   ABC093         H9  93(1,H9)    X93 plate1         27.0          46.67771
#> 82   ABC094        H10 94(1,H10)    X94 plate1         24.0          27.75898
#> 83   ABC095        H11 95(1,H11)    X95 plate1         34.5         108.78336
#> 84   ABC096        H12 96(1,H12)    X96 plate1         27.5          50.20976
#>    PvLDH-ST_DilutionReciprocal PvLDH-ST_MinStd PvLDH-ST_MaxDilution
#> 1                 3.016885e-03              30                50.81
#> 2                 1.737051e-02              30                50.81
#> 3                 1.004922e-02              30                50.81
#> 4                 1.161282e-02              30                50.81
#> 5                 1.052923e-02              30                50.81
#> 6                 1.443019e-02              30                50.81
#> 7                 5.934506e-06              30                50.81
#> 8                 6.100947e-03              30                50.81
#> 9                 4.238046e-03              30                50.81
#> 10                1.161282e-02              30                50.81
#> 11                1.004922e-02              30                50.81
#> 12                9.007870e-04              30                50.81
#> 13                1.628938e-02              30                50.81
#> 14                3.857325e-03              30                50.81
#> 15                5.728305e-03              30                50.81
#> 16                3.322336e-03              30                50.81
#> 17                1.628938e-02              30                50.81
#> 18                1.737051e-02              30                50.81
#> 19                6.986631e-03              30                50.81
#> 20                1.991645e-02              30                50.81
#> 21                6.100947e-03              30                50.81
#> 22                4.463669e-02              30                50.81
#> 23                1.161282e-02              30                50.81
#> 24                2.504711e-02              30                50.81
#> 25                1.104890e-02              30                50.81
#> 26                9.604842e-03              30                50.81
#> 27                2.724011e-02              30                50.81
#> 28                8.809348e-03              30                50.81
#> 29                1.741629e-03              30                50.81
#> 30                5.691249e-02              30                50.81
#> 31                1.737051e-02              30                50.81
#> 32                8.452413e-03              30                50.81
#> 33                3.602438e-02              30                50.81
#> 34                1.113144e-03              30                50.81
#> 35                2.504711e-02              30                50.81
#> 36                1.052923e-02              30                50.81
#> 37                2.724011e-02              30                50.81
#> 38                1.161282e-02              30                50.81
#> 39                1.471656e-03              30                50.81
#> 40                1.531376e-02              30                50.81
#> 41                2.975378e-02              30                50.81
#> 42                5.393682e-03              30                50.81
#> 43                4.238046e-03              30                50.81
#> 44                1.628938e-02              30                50.81
#> 45                3.602438e-02              30                50.81
#> 46                7.516615e-03              30                50.81
#> 47                7.243309e-03              30                50.81
#> 48                1.104890e-02              30                50.81
#> 49                2.724011e-02              30                50.81
#> 50                1.004922e-02              30                50.81
#> 51                1.289542e-02              30                50.81
#> 52                7.808068e-03              30                50.81
#> 53                5.691249e-02              30                50.81
#> 54                6.517854e-03              30                50.81
#> 55                7.243309e-03              30                50.81
#> 56                8.119366e-03              30                50.81
#> 57                5.691249e-02              30                50.81
#> 58                1.443019e-02              30                50.81
#> 59                1.161282e-02              30                50.81
#> 60                2.504711e-02              30                50.81
#> 61                2.724011e-02              30                50.81
#> 62                1.737051e-02              30                50.81
#> 63                1.289542e-02              30                50.81
#> 64                1.443019e-02              30                50.81
#> 65                2.142350e-02              30                50.81
#> 66                1.052923e-02              30                50.81
#> 67                1.991645e-02              30                50.81
#> 68                8.119366e-03              30                50.81
#> 69                5.691249e-02              30                50.81
#> 70                1.443019e-02              30                50.81
#> 71                3.265401e-02              30                50.81
#> 72                1.289542e-02              30                50.81
#> 73                2.312231e-02              30                50.81
#> 74                1.161282e-02              30                50.81
#> 75                2.504711e-02              30                50.81
#> 76                8.809348e-03              30                50.81
#> 77                3.602438e-02              30                50.81
#> 78                2.217579e-03              30                50.81
#> 79                1.004922e-02              30                50.81
#> 80                1.222629e-02              30                50.81
#> 81                2.142350e-02              30                50.81
#> 82                3.602438e-02              30                50.81
#> 83                9.192582e-03              30                50.81
#> 84                1.991645e-02              30                50.81
#>    PvLDH-ST_MaxStd PvLDH-ST_MinDilution
#> 1            23757                1e+06
#> 2            23757                1e+06
#> 3            23757                1e+06
#> 4            23757                1e+06
#> 5            23757                1e+06
#> 6            23757                1e+06
#> 7            23757                1e+06
#> 8            23757                1e+06
#> 9            23757                1e+06
#> 10           23757                1e+06
#> 11           23757                1e+06
#> 12           23757                1e+06
#> 13           23757                1e+06
#> 14           23757                1e+06
#> 15           23757                1e+06
#> 16           23757                1e+06
#> 17           23757                1e+06
#> 18           23757                1e+06
#> 19           23757                1e+06
#> 20           23757                1e+06
#> 21           23757                1e+06
#> 22           23757                1e+06
#> 23           23757                1e+06
#> 24           23757                1e+06
#> 25           23757                1e+06
#> 26           23757                1e+06
#> 27           23757                1e+06
#> 28           23757                1e+06
#> 29           23757                1e+06
#> 30           23757                1e+06
#> 31           23757                1e+06
#> 32           23757                1e+06
#> 33           23757                1e+06
#> 34           23757                1e+06
#> 35           23757                1e+06
#> 36           23757                1e+06
#> 37           23757                1e+06
#> 38           23757                1e+06
#> 39           23757                1e+06
#> 40           23757                1e+06
#> 41           23757                1e+06
#> 42           23757                1e+06
#> 43           23757                1e+06
#> 44           23757                1e+06
#> 45           23757                1e+06
#> 46           23757                1e+06
#> 47           23757                1e+06
#> 48           23757                1e+06
#> 49           23757                1e+06
#> 50           23757                1e+06
#> 51           23757                1e+06
#> 52           23757                1e+06
#> 53           23757                1e+06
#> 54           23757                1e+06
#> 55           23757                1e+06
#> 56           23757                1e+06
#> 57           23757                1e+06
#> 58           23757                1e+06
#> 59           23757                1e+06
#> 60           23757                1e+06
#> 61           23757                1e+06
#> 62           23757                1e+06
#> 63           23757                1e+06
#> 64           23757                1e+06
#> 65           23757                1e+06
#> 66           23757                1e+06
#> 67           23757                1e+06
#> 68           23757                1e+06
#> 69           23757                1e+06
#> 70           23757                1e+06
#> 71           23757                1e+06
#> 72           23757                1e+06
#> 73           23757                1e+06
#> 74           23757                1e+06
#> 75           23757                1e+06
#> 76           23757                1e+06
#> 77           23757                1e+06
#> 78           23757                1e+06
#> 79           23757                1e+06
#> 80           23757                1e+06
#> 81           23757                1e+06
#> 82           23757                1e+06
#> 83           23757                1e+06
#> 84           23757                1e+06
# }