Add specific attributes to a dataframe

Add specific attributes to a dataframe

Usage

add_attr(data, pathogen, adm0)

Arguments

data

A dataframe to which attributes will be added.

pathogen

Name of pathogen passed to an attribute of the same name. This will be retrieved from assays_df based on test_id provided by user.

adm0

Unique code of the country passed to an attribute of the same name. Code is based on serotrackr::regions object.

Value

A dataframe with additional attributes.

Note

An id attribute is also added using generate_obj_id().

Examples

df <- add_attr(sample_raw_data, pathogen = "SARS-CoV-2",
               adm0 = regions$adm0$Canada)
attributes(df)
#> $class
#> [1] "tbl_df"     "tbl"        "data.frame"
#> 
#> $row.names
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100
#> 
#> $names
#>  [1] "dataset_id"            "id"                    "age_group"            
#>  [4] "age"                   "sex"                   "adm0"                 
#>  [7] "adm1"                  "adm2"                  "collection_start_date"
#> [10] "collection_end_date"   "test_id"               "result"               
#> [13] "result_cat"            "another_col"          
#> 
#> $id
#> [1] 28711
#> 
#> $pathogen
#> [1] "SARS-CoV-2"
#> 
#> $adm0
#> [1] "40"
#>