Relabel column names to Standardised Naming Convention

This is a helper function to be used inside `readSeroData()` to relabel columns for each plate.

Usage

.relabel_columns(df)

Arguments

df

Data frame from `readSeroData()` processing.

Value

A data fame with columns renamed

Author

Dionne Argyropoulos

Examples

# \donttest{
your_raw_data <- system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR")

if (
 requireNamespace("dplyr", quietly = TRUE) &&
 requireNamespace("janitor", quietly = TRUE)
) {

  # Read in raw luminex file
  df <- .read_luminex_file(your_raw_data)

  # Get the start and end rows of the data section: start = "Median", end = "Net MFI"
  row1    <- which(df$xPONENT == "Median")
  row2    <- which(df$xPONENT == "Net MFI")

  # Apply data processing pipeline, including .relabel_columns()
  df |>
    dplyr::slice((row1 + 1):(row2 - 1)) |>
    janitor::row_to_names(row_number = 1) |>
    janitor::clean_names() |>
    dplyr::select(dplyr::where(~ !all(is.na(.x)))) |>
    dplyr::filter(dplyr::if_any(dplyr::everything(), ~ !is.na(.x))) |>
    dplyr::mutate(dplyr::across(everything(), ~ gsub("NaN", 0, .))) |>
    .relabel_columns()
}
#> Warning: Row 1 does not provide unique names. Consider running clean_names() after row_to_names().
#>     location     sample     EBP   LF005   LF010   LF016    MSP8 RBP2b.P87
#> 1    1(1,A1)     Blank1      20     200      20      10      20        10
#> 2    2(1,A2)     Blank2      15     291      15      15      15        10
#> 3    3(1,A3)         S1 15710.5   11990   22583   21244   24306    7907.5
#> 4    4(1,A4)         S2   13545    8285 16947.5   16146   17172      7186
#> 5    5(1,A5)         S3    9767    4950   10865 10621.5   11358    4475.5
#> 6    6(1,A6)         S4  5648.5    2519    6060  5968.5    6237      3508
#> 7    7(1,A7)         S5  4104.5    1431  3711.5    3738  3883.5      2082
#> 8    8(1,A8)         S6    2105   676.5    1889    1667    2136      1102
#> 9    9(1,A9)         S7    1107     328    1106     890    1204       657
#> 10 10(1,A10)         S8   452.5     141     465   405.5     522       277
#> 11 11(1,A11)         S9     209      65   206.5   181.5     249       138
#> 12 12(1,A12)        S10     105      42   113.5      87     134        93
#> 13  13(1,B1) Unknown013    2712    1569     673     327     182      1068
#> 14  14(1,B2) Unknown014     134     378     117     197      58     465.5
#> 15  15(1,B3) Unknown015     182     209     208     374   221.5       463
#> 16  16(1,B4) Unknown016     152   229.5     101      89      48       591
#> 17  17(1,B5) Unknown017    1135     236     299     507   209.5       671
#> 18  18(1,B6) Unknown018     174     395     175      78      70       103
#> 19  19(1,B7) Unknown019     421  2081.5     529     149     962    1441.5
#> 20  20(1,B8) Unknown020      24      49      22      13    15.5        14
#> 21  21(1,B9) Unknown021      24      45      22      13      16        13
#> 22 22(1,B10) Unknown022   21464   11789  2508.5    8001    2342      1386
#> 23 23(1,B11) Unknown023     795  6135.5     407     273   535.5      1575
#> 24 24(1,B12) Unknown024    1574   348.5     892     288     306       590
#> 25  25(1,C1) Unknown025     146   408.5  1130.5      83     100       301
#> 26  26(1,C2) Unknown026  1330.5    3036     965     174      79      1090
#> 27  27(1,C3) Unknown027     358  4074.5   335.5     140     390      6704
#> 28  28(1,C4) Unknown028     481    1870     421   182.5  1777.5      2390
#> 29  29(1,C5) Unknown029     551    1282   331.5     165      77       244
#> 30  30(1,C6) Unknown030     943   584.5     235     135    3017      1034
#> 31  31(1,C7) Unknown031   532.5    1533     315     210  2575.5      2903
#> 32  32(1,C8) Unknown032     768   12605     409    7632    9860     13703
#> 33  33(1,C9) Unknown033     130   239.5      72      64      37       156
#> 34 34(1,C10) Unknown034      40      60      37      25    26.5      23.5
#> 35 35(1,C11) Unknown035     167     605     253     329    3973    1426.5
#> 36 36(1,C12) Unknown036    1879    4562     415   18287 11482.5     12031
#> 37  37(1,D1) Unknown037     256    1507     117     129     168       488
#> 38  38(1,D2) Unknown038      24      49      22      12      15        14
#> 39  39(1,D3) Unknown039     384    1708     326  1172.5     986      1041
#> 40  40(1,D4) Unknown040     210     512     221     191      89       203
#> 41  41(1,D5) Unknown041   351.5     312   165.5      88      66       205
#> 42  42(1,D6) Unknown042    1278    5810     275   230.5     137      2032
#> 43  43(1,D7) Unknown043     318     397     130   130.5      78       903
#> 44  44(1,D8) Unknown044   493.5  2980.5     401  5932.5    3536    4699.5
#> 45  45(1,D9) Unknown045    9686     873     333     166     791      1093
#> 46 46(1,D10) Unknown046   777.5     581   241.5     173  1479.5       899
#> 47 47(1,D11) Unknown047     455   416.5     126      67      48       400
#> 48 48(1,D12) Unknown048     202    1142   137.5     176   204.5      1090
#> 49  49(1,E1) Unknown049   333.5   623.5     368   10430    2340      4375
#> 50  50(1,E2) Unknown050     561    7984     441     128     141      1993
#> 51  51(1,E3) Unknown051   506.5     340   152.5     147      90       600
#> 52  52(1,E4) Unknown052     254     367   392.5      89   142.5       280
#> 53  53(1,E5) Unknown053     888    1955     438     173     507      1009
#> 54  54(1,E6) Unknown054     426    1877     186    9767   15945     10073
#> 55  55(1,E7) Unknown055    5733  1301.5  1920.5   220.5    4833      2340
#> 56  56(1,E8) Unknown056   19092   14422    1993 12836.5    7426     11930
#> 57  57(1,E9) Unknown057     596   15139   371.5     229   535.5     14032
#> 58 58(1,E10) Unknown058    1028    2646    1486     292    9357      8009
#> 59 59(1,E11) Unknown059     512    4735    3408   127.5     191      4401
#> 60 60(1,E12) Unknown060     883    1953     755     586     122       830
#> 61  61(1,F1) Unknown061     977    8719     509     187    1289     12930
#> 62  62(1,F2) Unknown062     791    1944     465   216.5   12307      5029
#> 63  63(1,F3) Unknown063     397     987     924     133      79       253
#> 64  64(1,F4) Unknown064    1111     613     357     196    8496      7914
#> 65  65(1,F5) Unknown065     956    2369     644     389    8762      9649
#> 66  66(1,F6) Unknown066    1307   20590    1335   10780   12736   25375.5
#> 67  67(1,F7) Unknown067     433     382     173     105      59       480
#> 68  68(1,F8) Unknown068     161     279  2524.5   105.5      55       307
#> 69  69(1,F9) Unknown069     439   732.5     397    2108    9692      8090
#> 70 70(1,F10) Unknown070    1837   11406   467.5   20011   19606   24476.5
#> 71 71(1,F11) Unknown071     244  1665.5     166     124     154       224
#> 72 72(1,F12) Unknown072     482  3990.5     388    3736    7436      7361
#> 73  73(1,G1) Unknown073     656    3103     505   14386    2343      5503
#> 74  74(1,G2) Unknown074     243   568.5     693     141      79       339
#> 75  75(1,G3) Unknown075  1409.5    1123     379     153   167.5      1013
#> 76  76(1,G4) Unknown076   467.5    6741     238     174     181      6160
#> 77  77(1,G5) Unknown077     329     630     196     119     278       787
#> 78  78(1,G6) Unknown078  1122.5   17810     677   12254   16011     18719
#> 79  79(1,G7) Unknown079   10853   817.5     255     189  2247.5      3322
#> 80  80(1,G8) Unknown080   743.5     820   313.5     289  9351.5      7171
#> 81  81(1,G9) Unknown081     723    1255     299     253      95       712
#> 82 82(1,G10) Unknown082     168     684     131     198   135.5       392
#> 83 83(1,G11) Unknown083     180     276     205  4646.5    2550    2664.5
#> 84 84(1,G12) Unknown084   249.5    4104     250      96      68      1020
#> 85  85(1,H1) Unknown085     479     190     105     131      66      1449
#> 86  86(1,H2) Unknown086     164     327   234.5      66    75.5       259
#> 87  87(1,H3) Unknown087   281.5    1794     413      75   469.5       802
#> 88  88(1,H4) Unknown088     269    1067     141    7595   14464      7592
#> 89  89(1,H5) Unknown089    3247     382     470      98  2319.5      2009
#> 90  90(1,H6) Unknown090   20593 14902.5    2021   10184  5805.5   14083.5
#> 91  91(1,H7) Unknown091     395    9914     211   131.5     325      3021
#> 92  92(1,H8) Unknown092   904.5    1871     793     172    6106      2093
#> 93  93(1,H9) Unknown093     323  2007.5    1950     106     191      2411
#> 94 94(1,H10) Unknown094     706    2087     791     169      66       990
#> 95 95(1,H11) Unknown095     323    4950     264     107     645      7092
#> 96 96(1,H12) Unknown096     367    1067     259     101    7423      1203
#> 97                                                                       
#>    PTEX150   PvCSS total_events x x_2 x_3 x_4 x_5 x_6
#> 1       30      20         1898                      
#> 2       20      15         1805                      
#> 3    13207   12427         1842                      
#> 4     8550    8034         2044                      
#> 5     5329    4990         1815                      
#> 6     2671  2446.5         1950                      
#> 7     1548    1299         1991                      
#> 8    698.5     581         1827                      
#> 9      333     264         2095                      
#> 10     156     111         1643                      
#> 11      77    46.5         1888                      
#> 12      40      28         1876                      
#> 13     936     223         1927                      
#> 14     122      93         1965                      
#> 15     293     868         1970                      
#> 16     109     110         1911                      
#> 17  1665.5     266         1977                      
#> 18   294.5      92         1902                      
#> 19     241     282         1913                      
#> 20      28      17         1828                      
#> 21      29      17         1885                      
#> 22    1992    2927         2133                      
#> 23    8821     998         1670                      
#> 24     315     366         1955                      
#> 25     652     130         1873                      
#> 26     179      92         2173                      
#> 27    8814    5627         2012                      
#> 28    2465     411         1794                      
#> 29     579     229         2023                      
#> 30    4928    7730         1893                      
#> 31    2779    7300         1946                      
#> 32    5612    6383         1843                      
#> 33     204      49         1900                      
#> 34      40      27         1947                      
#> 35     803    1670         1826                      
#> 36  4830.5    4217         1690                      
#> 37     908      87         1988                      
#> 38      29      17         1943                      
#> 39   299.5     149         1963                      
#> 40     410   174.5         2159                      
#> 41     200     183         2122                      
#> 42    1762    1241         1997                      
#> 43  1513.5     325         1916                      
#> 44     574    5389         2026                      
#> 45   417.5     262         1691                      
#> 46    1504     260         2044                      
#> 47     183      61         1639                      
#> 48     504     459         1957                      
#> 49    1650  7801.5         1836                      
#> 50    2147    1502         1963                      
#> 51    2178     170         1856                      
#> 52     298     138         2019                      
#> 53     709    2040         2009                      
#> 54  1585.5    1124         1964                      
#> 55    5329    2369         1905                      
#> 56    4826  3468.5         2052                      
#> 57  7928.5  1414.5         1712                      
#> 58    3736 14370.5         1862                      
#> 59   11567  2923.5         1718                      
#> 60   425.5     121         1767                      
#> 61   13891    9531         1735                      
#> 62    4157     760         1804                      
#> 63  2317.5  1080.5         1705                      
#> 64  5772.5   13202         2026                      
#> 65    6437   21113         1871                      
#> 66    5493    7696         1868                      
#> 67     655     367         1813                      
#> 68   681.5   472.5         2070                      
#> 69    9513    7341         1870                      
#> 70    5822  5127.5         2055                      
#> 71     718     115         1798                      
#> 72  4111.5    2819         1831                      
#> 73   357.5    1215         1816                      
#> 74    2080    1470         1821                      
#> 75    1553    1738         1979                      
#> 76  1660.5    1087         1910                      
#> 77  2052.5    2221         1752                      
#> 78    3551 10222.5         1670                      
#> 79     590     258         1702                      
#> 80    3062   676.5         1739                      
#> 81     396   129.5         1574                      
#> 82     292   254.5         1597                      
#> 83     724    2707         1546                      
#> 84     960     672         1710                      
#> 85  1724.5     108         2102                      
#> 86     260      89         1895                      
#> 87     345   744.5         1694                      
#> 88   804.5     467         1817                      
#> 89    1521    1258         1615                      
#> 90  4307.5    3073         1452                      
#> 91    4608  1073.5         1537                      
#> 92    1303  6595.5         1852                      
#> 93  7377.5    1699         1563                      
#> 94     261      65         1472                      
#> 95    8856    5330         1467                      
#> 96  1841.5     335         1654                      
#> 97                                                   
# }