Map columns of data

Maps the required columns in your data and validates them.

Usage

map_cols(
  data,
  dataset_id = NULL,
  id = NULL,
  age_group,
  sex,
  country,
  state = NULL,
  county = NULL,
  city = NULL,
  collection_start_date,
  collection_end_date,
  test_id,
  result,
  result_cat,
  include_others = FALSE
)

Arguments

data

a data.frame

dataset_id

column, data collection event

id

column for IDs

age_group

column containing age min and age max for each age group

sex

column for sex

country

a string, name of the country

state

column for state

county

column for county

city

column for city

collection_start_date

column for sampling start date

collection_end_date

column for sampling end date

test_id

Will add the list of test later

result

column with test results

result_cat

column with values of "positive", "negative", or "borderline"

include_others

include additional columns or not

Value

A cleaned data.frame

Examples

library(dplyr)
library(stringr)
df_orig <- data.frame(dataset_id = 1,
                      id = 1:3,
                      age = c("0 - 10", "20-30", "30,40"),
                      sex = c("f", "Male", "binary"),
                      collection_start_date = "2020-03-01",
                      collection_end_date = "2020-04-01",
                      result = c("'2.4'", "3.5", "9"),
                      result_cat = rep("negative", times = 3))

map_cols(df_orig,
         dataset_id = dataset_id,
         id = id,
         age_group = age,
         sex = sex,
         country = "Earth!",
         collection_start_date = collection_start_date,
         collection_end_date = collection_end_date,
         test_id = assays$`SARS-CoV-2`$`EUROIMMUN - IgG - Anti-SARS-CoV-2 ELISA IgG`,
         result = result,
         result_cat = result_cat,
         include_others = FALSE)
#>   dataset_id id age_group    sex country state county city
#> 1          1  1    0 - 10      f  Earth!    NA     NA   NA
#> 2          1  2     20-30   Male  Earth!    NA     NA   NA
#> 3          1  3     30,40 binary  Earth!    NA     NA   NA
#>   collection_start_date collection_end_date
#> 1            2020-03-01          2020-04-01
#> 2            2020-03-01          2020-04-01
#> 3            2020-03-01          2020-04-01
#>                                       test_id result result_cat
#> 1 EUROIMMUN - IgG - Anti-SARS-CoV-2 ELISA IgG  '2.4'   negative
#> 2 EUROIMMUN - IgG - Anti-SARS-CoV-2 ELISA IgG    3.5   negative
#> 3 EUROIMMUN - IgG - Anti-SARS-CoV-2 ELISA IgG      9   negative