Message when supplied type is wrong

Message when supplied type is wrong

Usage

msg_wrongType(
  arg,
  req_type = c("char", "num", "date_char"),
  mode = c("col", "val", "no_typeof"),
  name = NULL,
  data = NULL,
  ...
)

Arguments

arg

Preset column name or argument for which the validation is run.

req_type

Type that was expected for this arg.

mode

Defines if validation is done for a column (col) or for a single value scalar (val). This will determine how the current wrong type of the supplied column or vector should be evaluated. The no_typeof option is for when we do not need the message to print the type of the supplied input.

name

Only for modes col and val. The column name, supplied as deparse(substitute(argument)) or input value as substitute(argument).

data

A dataframe. Only for mode col.

...

Additional argument passed to msg_result().

Value

Prints a message to the console.

Examples

msg_wrongType("age", req_type = "num", mode = "no_typeof")
#> 
#> ✖ age must be an unquoted name of a numeric column or a number scalar.