Run Pk/Pf/Pv Data Analysis Pipeline from Start to End

Run Pk/Pf/Pv Data Analysis Pipeline from Start to End

Usage

runPlasmoPipeline(
  raw_data,
  platform = "magpix",
  plate_layout,
  panel = "panel1",
  std_point,
  experiment_name = "experiment1",
  classify = "Yes",
  algorithm_type = "antibody_model",
  sens_spec = "balanced"
)

Arguments

raw_data

String with the raw data path.

platform

"magpix" or "bioplex". Default: "Bioplex"

plate_layout

An ".xlsx" file with sheets labelled plate1, plate2... etc.

panel

Panel of Pk/Pf/Pv antigens. Default = "panel1" or user provided csv of Antigens and Species.

std_point

Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve. Value is an integer.

experiment_name

User-input experiment name. Default: "experiment1".

classify

"Yes" or "No" depending on whether you would like classification or not. Default = "Yes".

algorithm_type

Algorithm: "antibody_model" (PvSEM algorithm; default)

sens_spec

User-selected Sensitivity/Specificity threshold: "balanced" (default) or "90% specificity".

Value

A data frame containing the MFI and RAU Dilution values for each sample, QC plots for standard curve, bead counts and blanks.

Author

Dionne Argyropoulos

Examples

# \donttest{

# Helper to avoid repetition in examples
run_example_std <- function(std_point) {
  # Load raw data for given standard curve
  your_raw_data <- c(
    system.file("extdata",
                paste0("example_MAGPIX_pk_", std_point, "std_plate1.csv"),
                package = "SeroTrackR"),
    system.file("extdata",
                paste0("example_MAGPIX_pk_", std_point, "std_plate2.csv"),
                package = "SeroTrackR")
  )

  layout_file <- system.file(
    "extdata",
    paste0("example_platelayout_pk_", std_point, "std.xlsx"),
    package = "SeroTrackR"
  )

  # Run pipeline
  runPlasmoPipeline(
    raw_data = your_raw_data,
    platform = "magpix",
    plate_layout = layout_file,
    panel = "panel1",
    std_point = std_point,
    experiment_name = paste0(std_point, "-point standard curve")
  )
}

# ---- 5-point standard curve ----
results_5std <- run_example_std(5)
#> PASS: File example_magpix_pk_5std_plate1.csv successfully validated.
#> PASS: File example_magpix_pk_5std_plate2.csv successfully validated.
#> Plate layouts correctly identified!
#> QC Processes completed.
#> QC Plotting completed.
#> MFI to RAU conversion completed.
#> Pv classification completed.

# ---- 10-point standard curve ----
results_10std <- run_example_std(10)
#> PASS: File example_magpix_pk_10std_plate1.csv successfully validated.
#> PASS: File example_magpix_pk_10std_plate2.csv successfully validated.
#> Plate layouts correctly identified!
#> QC Processes completed.
#> QC Plotting completed.
#> MFI to RAU conversion completed.
#> Pv classification completed.

# }