Aggregate data

Generate aggregate estimates based on validated individual level data.

Usage

st_aggregate(
  data,
  subgroup = c("age_group", "sex", "age_group + sex"),
  borderline = c("negative", "positive", NA),
  add_ci = TRUE,
  round_digits = 4,
  test_combination = NULL
)

Arguments

data

A validated data.frame that is the output of st_validate()

subgroup

A character vector of subgrouping variables. By default, aggregate estimates are generated for the overall data, as well as age group, sex, and age group + sex subgroups.

borderline

How should borderline results be treated? Default is as negative.

add_ci

Boolean. Whether to add binomial proportion confidence interval. It is calculated using the Wilson score interval method through the binom::binom.confint() function.

round_digits

Integer indicating the number of decimal places of the estimate. It is passed to the digits argument of base::round().

test_combination

Not functional yet. When data is based on more than one assay, what is the relationship between those assays?

Value

A summarized data.frame

Examples

new_raw_data <- dplyr::mutate(
  sample_raw_data,
  age_group = rep(c("0-9", "10-19", "20-29", "30-39", "40+"), each=20),
  age = c(sample(0:9, 20, replace=TRUE), sample(10:19, 20, replace=TRUE),
          sample(20:29, 20, replace=TRUE), sample(30:39, 20, replace=TRUE),
          sample(40:120, 20, replace=TRUE)),
  sex = c(rep("f", 40), rep("m", 40), rep("o", 20)),
  result_cat = dplyr::case_when(result_cat == "neg" ~ "negative",
                                result_cat == "pos" ~ "positive",
                                TRUE ~ result_cat)
)

validated_df <- st_validate(
  new_raw_data,
  dataset_id = dataset_id,
  id = id,
  age_group = age_group,
  age = age,
  sex = sex,
  adm0 = regions$adm0$Canada,
  adm1 = regions$adm1$Canada$Alberta,
  adm2 = regions$adm2$Canada$Alberta$Calgary,
  collection_start_date = "2023-01-01",
  collection_end_date = "2023-02-01",
  test_id = assays$`SARS-CoV-2`$`AAZ LMB - IgG, IgM - COVID-PRESTO®`,
  result = result,
  result_cat = result_cat,
  include_others = TRUE
)
#> 
#> ── Mapping columns and validating data ─────────────────────────────────────────
#> ℹ age_group ...
#> ✔ age_group is a valid column. [476ms]
#> ℹ age ...
#> ✔ age is a valid column. [21ms]
#> ℹ sex ...
#> ✔ sex is a valid column. [11ms]
#> ℹ adm0 ...
#> ✔ adm0 is a valid string. [96ms]
#> ℹ adm1 ...
#> ✔ adm1 is a valid string. [9ms]
#> ℹ adm2 ...
#> ✔ adm2 is a valid string. [14ms]
#> ℹ collection_start_date ...
#> ✔ collection_start_date is a valid scalar. [166ms]
#> ℹ collection_end_date ...
#> ✔ collection_end_date is a valid scalar. [18ms]
#> ℹ test_id ...
#> ✔ test_id is a valid string. [7ms]
#> ℹ result ...
#> ✔ result is a valid column. [9ms]
#> ℹ result_cat ...
#> ✔ result_cat is a valid column. [8ms]
#> ℹ dataset_id ...
#> ✔ dataset_id is a valid column. [3ms]
#> ℹ id ...
#> ✔ id is a valid column. [11ms]
#> 
#> ── Validation finished ─────────────────────────────────────────────────────────
#> Success! Validated data created.
#> 

st_aggregate(validated_df)
#> # A tibble: 28 × 27
#>    dataset_id subgroup  strata age_group age_min age_max sex   pop_adj test_adj
#>         <dbl> <chr>     <chr>  <chr>       <int>   <int> <chr> <lgl>   <lgl>   
#>  1          1 overall   NA     All             1     119 All   FALSE   FALSE   
#>  2          2 overall   NA     All             1     107 All   FALSE   FALSE   
#>  3          1 age_group 0-9    0-9             1       7 All   FALSE   FALSE   
#>  4          1 age_group 10-19  10-19          10      19 All   FALSE   FALSE   
#>  5          1 age_group 20-29  20-29          20      29 All   FALSE   FALSE   
#>  6          1 age_group 30-39  30-39          30      38 All   FALSE   FALSE   
#>  7          1 age_group 40+    40+            47     119 All   FALSE   FALSE   
#>  8          2 age_group 0-9    0-9             1       9 All   FALSE   FALSE   
#>  9          2 age_group 10-19  10-19          11      19 All   FALSE   FALSE   
#> 10          2 age_group 20-29  20-29          20      29 All   FALSE   FALSE   
#> # ℹ 18 more rows
#> # ℹ 18 more variables: adm1 <chr>, adm2 <chr>, start_date <date>,
#> #   end_date <date>, test_id_1 <chr>, test_id_2 <chr>, test_id_3 <chr>,
#> #   test_combination <lgl>, numerator <dbl>, denominator <int>, seroprev <dbl>,
#> #   seroprev_95_ci_lower <dbl>, seroprev_95_ci_upper <dbl>,
#> #   ab_denominator <int>, ab_titer_min <dbl>, ab_titer_max <dbl>,
#> #   ab_titer_mean <dbl>, ab_titer_sd <dbl>